+++ src/layout/RootLayout.tsx
@@ -1,5 +1,5 @@
 import '../index.css';
-import Navigation from './Navigation.tsx';
+import Navigation from './Navbar';
 import Footer from './Footer';
 import {Outlet} from 'react-router-dom';
 // container max-w-[1440px] mx-auto
